스폰볼륨 생성

+++ Source/UnrealPractice3/Public/MineItem.h
@@ -21,7 +21,7 @@ protected:
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Mine")
     float ExplosionRadius;
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Mine")
-    float ExplosionDamage;
+    int32 ExplosionDamage;
 
     FTimerHandle ExplosionTimerHandle;
